Split index tests in separate file
